Support YAML File testing more wholistically.
