+++ tripal/tests/src/Traits/TripalTestTrait.php
@@ -8,6 +8,7 @@ use Drupal\tripal\TripalVocabTerms\Interfaces\TripalVocabularyInterface;
 use Drupal\tripal\TripalVocabTerms\Interfaces\TripalIdSpaceInterface;
 use Drupal\tripal\TripalVocabTerms\TripalTerm;
 use PHPUnit\Framework\Attributes\Group;
+use Symfony\Component\Yaml\Yaml;
 
 /**
  * Provides functions related to setting up Tripal test environments
@@ -403,6 +404,80 @@ trait TripalTestTrait {
     }
     $modules_array_code = 'protected static $modules = [\'' . implode("','", $suggested_modules) . '\'];';
     $this->assertEmpty($missing_modules, 'You are missing some modules in your static $modules array. For the functionality you requested, we suggest the following: ' . $modules_array_code);
+  }
+
+  /**
+   * Retrieves test information from a YAML file.
+   *
+   * @param string $yaml_file
+   *   The full path to a yaml file which follows the format descripbed above.
+   *
+   * @return array
+   *   The first array returned describes the state of the test environment to
+   *   be setup and the second describes the scenarios to test. For a
+   *   description of the structure of these arrays, see the YAML file directly.
+   */
+  public function getTestInfoFromYaml(string $yaml_file): array {
+
+    if (!file_exists($yaml_file)) {
+      throw new \Exception("Cannot open YAML file $yaml_file.");
+    }
+
+    $file_contents = file_get_contents($yaml_file);
+    if (empty($file_contents)) {
+      throw new \Exception("Unable to retrieve contents for YAML file $yaml_file.");
+    }
+
+    $yaml_data = Yaml::parse($file_contents);
+    if (empty($yaml_data)) {
+      throw new \Exception("Unable to parse YAML file $yaml_file.");
+    }
+
+    if (!array_key_exists('system-under-test', $yaml_data)) {
+      throw new \Exception("The 'system-under-test' key is missing from the $yaml_file.");
+    }
+    if (!array_key_exists('scenarios', $yaml_data)) {
+      throw new \Exception("The 'scenarios' key is missing from the $yaml_file.");
+    }
+
+    return [$yaml_data['system-under-test'], $yaml_data['scenarios']];
+  }
+
+  /**
+   * Retrieves a specific scenario defined in a YAML file.
+   *
+   * If the YAML was preprocessed then the scenarios will be pulled
+   * directly from $this->scenarios. Otherwise, it the YAML saved in
+   * $this->yaml_info_file will first be processed.
+   *
+   * @param int $scenario_key
+   *   The key of the scenario we want to get.
+   * @param string $scenario_label
+   *   The label of the scenario we want to get.
+   *
+   * @return array
+   *   The specific scenario array.
+   */
+  public function getYamlScenario(int $scenario_key, string $scenario_label): array {
+
+    // Process the YAML test file if it hasn't already been done.
+    if (!is_array($this->scenarios) or empty($this->scenarios)) {
+      $this->assertFileIsReadable($this->yaml_info_file, 'The YAML file defining the test (see $this->yaml_info_file) was either not set or is not readable.');
+      [$this->system_under_test, $this->scenarios] = $this->getTestInfoFromYaml($this->yaml_info_file);
+    }
+
+    // Confirm the scenario key exists.
+    $this->assertArrayHasKey($scenario_key, $this->scenarios, "We tried to retrieve the '$scenario_label' (key: $scenario_key) but it didn't exist.");
+
+    // Retrieve the scenario.
+    $scenario = $this->scenarios[$scenario_key];
+
+    // Confirm the label matches our expectations.
+    $this->assertEquals($scenario_label, $scenario['label'], "We tried to retrieve the '$scenario_label' (key: $scenario_key) but its label did not match our expectations.");
+
+    // @todo support processing the scenario (e.g. set type_ids)
+    //
+    return $scenario;
   }
 
 }
