+++ tripal_chado/tests/src/Kernel/Plugin/ChadoField/FieldType/ChadoAdditionalTypeTest.php
@@ -165,8 +165,7 @@ class ChadoAdditionalTypeTest extends ChadoTestKernelBase {
   public function testFieldTypeCrud(int $current_scenario_key, string $current_scenario_label) {
 
     // Retrieve the correct scenario.
-    $current_scenario = $this->scenarios[$current_scenario_key];
-    $this->assertEquals($current_scenario_label, $current_scenario['label'], "We may not have retrieved the expected scenario as the labels did not match.");
+    $current_scenario = $this->getYamlScenario($current_scenario_key, $current_scenario_label);
 
     // 1. Create the entity with that value set.
     $entity = TripalEntity::create([
